Refactor localStorage data handling; move Vertex AI logic to vertexAI.js and initialize on DOMContentLoaded
