+++ index.html
@@ -6,10 +6,7 @@
     <title>Weather Dashboard</title>
     <link rel="stylesheet" href="CurrentAndHourly.css">
 </head>
-<script type="module">
-    import { localStorageDataChecks } from "./js/app.js";
-    localStorageDataChecks();
-</script>
+<script type="module" src="./js/vertexAI.js"></script>
 <body>
     <div class="container">
         <div class="header">
